requirements-dev.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile requirements-dev.in
#
appdirs==1.4.4
    # via black
asgiref==3.4.1
    # via django
attrs==21.2.0
    # via pytest
black==21.7b0
    # via -r requirements-dev.in
boto3==1.18.29
    # via django-s3-folder-storage
botocore==1.21.29
    # via
    #   boto3
    #   s3transfer
certifi==2021.5.30
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.0.4
    # via requests
click==8.0.1
    # via black
codecov==2.1.12
    # via -r requirements-dev.in
collectfast==2.2.0
    # via -r requirements.in
coverage==5.5
    # via
    #   codecov
    #   pytest-cov
dj-database-url==0.5.0
    # via -r requirements.in
django==3.2.6
    # via
    #   -r requirements.in
    #   collectfast
    #   django-debug-toolbar
    #   django-storages
    #   model-bakery
django-debug-toolbar==3.2.2
    # via -r requirements-dev.in
django-ordered-model==3.4.3
    # via -r requirements.in
django-s3-folder-storage==0.5
    # via -r requirements.in
django-storages==1.11.1
    # via
    #   collectfast
    #   django-s3-folder-storage
flake8==3.9.2
    # via -r requirements-dev.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.2
    # via requests
iniconfig==1.1.1
    # via pytest
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
mccabe==0.6.1
    # via flake8
model-bakery==1.3.2
    # via -r requirements-dev.in
mypy-extensions==0.4.3
    # via black
packaging==21.0
    # via pytest
pathspec==0.9.0
    # via black
pluggy==0.13.1
    # via pytest
psycopg2-binary==2.9.1
    # via -r requirements.in
py==1.10.0
    # via pytest
pycodestyle==2.7.0
    # via flake8
pyflakes==2.3.1
    # via flake8
pyparsing==2.4.7
    # via packaging
pytest==6.2.4
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==2.12.1
    # via -r requirements-dev.in
pytest-django==4.4.0
    # via -r requirements-dev.in
python-dateutil==2.8.2
    # via botocore
python-decouple==3.4
    # via -r requirements.in
pytz==2021.1
    # via django
regex==2021.8.21
    # via black
requests==2.26.0
    # via codecov
s3transfer==0.5.0
    # via boto3
sentry-sdk==1.3.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.1
    # via
    #   django
    #   django-debug-toolbar
toml==0.10.2
    # via
    #   pytest
    #   pytest-cov
tomli==1.2.1
    # via black
typing-extensions==3.10.0.0
    # via collectfast
urllib3==1.26.6
    # via
    #   botocore
    #   requests
    #   sentry-sdk

# The following packages are considered to be unsafe in a requirements file:
# setuptools
